LPD-86378 Split again both functions into two separated files, because one is intended to be used inside the submodule, and the other one from outside, any other solution ends up in cross-submodule error
