+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/api/FDSConnection.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {getFDSAtom, getOrCreateSelector} from '../utils/getFDSAtom';
+import {getFDSAtom, getOrCreateSelector} from './getFDSAtom';
 
 import type {
 	FDSConnectionInfo,
