+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/index.ts
@@ -45,7 +45,7 @@ export {readConfigFromURL} from './utils/configInURL';
 export {getConfigParamName, serializeFDSConfig} from './utils/configInURL';
 export {default as FDS_EVENT} from './utils/eventsDefinitions';
 
-export {getFDSAtom, getOrCreateFDSAtom} from './utils/getFDSAtom';
+export {getOrCreateFDSAtom} from './utils/getOrCreateFDSAtom';
 
 export {
 	DisplayType,
